Tighten connection diagnostic module boundaries
