+++ apps/mobile/src/lib/connection-diagnostic-recorder.ts
@@ -1,6 +1,6 @@
+import { createConnectionDiagnosticEvent } from './connection-diagnostic-normalization';
 import {
 	cloneDiagnosticValue,
-	createConnectionDiagnosticEvent,
 	redactDiagnosticText,
 } from './connection-diagnostic-redaction';
 import {
