+++ apps/mobile/src/lib/connection-diagnostics.ts
@@ -1,9 +1,5 @@
 export * from './connection-diagnostic-types';
 export {
-	cloneDiagnosticValue,
-	normalizeConnectionIdentity,
-	normalizeTraceForPrompt,
-	redactDiagnosticText,
 	serializeConnectionDiagnosticError,
 } from './connection-diagnostic-redaction';
 export {
